requirements-prod.txt
# Production Python dependencies
gunicorn==21.2.0
gevent==23.7.0
redis==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
prometheus-client==0.17.1
flask-limiter==3.5.0
flask-cors==4.0.0
cryptography==41.0.4
requests==2.31.0
